Factor out some pre-Python runtime support and implement PYTHONINSPECT.
This change implements Py.getenv for access to os.environ, and uses it to
so that PYTHONINSPECT may trigger an interactive session sfter a script.
At the same time, we separate out some utilities that work before the Python
type system is ready.

--HG--
extra : amend_source : 8b00e2d41a9b748ac24b86588e610de352c30a12
