+++ src/org/python/core/Py.java
@@ -16,12 +16,6 @@ import java.io.Serializable;
 import java.io.StreamCorruptedException;
 import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
-import java.net.JarURLConnection;
-import java.net.MalformedURLException;
-import java.net.URI;
-import java.net.URISyntaxException;
-import java.net.URL;
-import java.net.URLConnection;
 import java.sql.Date;
 import java.sql.Time;
 import java.sql.Timestamp;
@@ -43,7 +37,7 @@ import jnr.posix.POSIX;
 import jnr.posix.POSIXFactory;
 import jnr.posix.util.Platform;
 
-public final class Py {
+public final class Py extends PrePy {
 
     static class SingletonResolver implements Serializable {
 
@@ -788,6 +782,71 @@ public final class Py {
             list.add(fileSystemDecode(filename));
         }
         return list;
+    }
+
+    /**
+     * Get the environment variables from {@code os.environ}. Keys and values should be
+     * {@code PyString}s in the file system encoding, and it may be a {@code dict} but nothing can
+     * be guaranteed. (Note that in the case of multiple interpreters, the target is in the current
+     * interpreter's copy of {@code os}.)
+     *
+     * @return {@code os.environ}
+     */
+    private static PyObject getEnvironment() {
+        PyObject os = imp.importName("os", true);
+        PyObject environ = os.__getattr__("environ");
+        return environ;
+    }
+
+    /** The same as {@code getenv(name, null)}. See {@link #getenv(PyString, PyString)}. */
+    public static PyString getenv(PyString name) {
+        return getenv(name, null);
+    }
+
+    /**
+     * Get the value of the environment variable named from {@code os.environ} or return the given
+     * default value. Empty string values are treated as undefined for this purpose.
+     *
+     * @param name of the environment variable.
+     * @param defaultValue to return if {@code key} is not defined (may be {@code null}.
+     * @return the corresponding value or <code>defaultValue</code>.
+     */
+    public static PyString getenv(PyString name, PyString defaultValue) {
+        try {
+            PyObject value = getEnvironment().__finditem__(name);
+            if (value == null) {
+                return defaultValue;
+            } else {
+                return value.__str__();
+            }
+        } catch (PyException e) {
+            // Something is fishy about os.environ, so the name is not defined.
+            return defaultValue;
+        }
+    }
+
+    /** The same as {@code getenv(name, null)}. See {@link #getenv(String, String)}. */
+    public static String getenv(String name) {
+        return getenv(name, null);
+    }
+
+    /**
+     * Get the value of the environment variable named from {@code os.environ} or return the given
+     * default value. This is a convenience wrapper on {@link #getenv(PyString, PyString)} which
+     * takes care of the fact that environment variables are FS-encoded.
+     *
+     * @param name to access in the environment.
+     * @param defaultValue to return if {@code key} is not defined.
+     * @return the corresponding value or <code>defaultValue</code>.
+     */
+    public static String getenv(String name, String defaultValue) {
+        PyString value = getenv(newUnicode(name), null);
+        if (value == null) {
+            return defaultValue;
+        } else {
+            // Environment variables are FS-encoded byte strings
+            return fileSystemDecode(value);
+        }
     }
 
     public static PyStringMap newStringMap() {
@@ -2566,155 +2625,6 @@ public final class Py {
             objs.add(item);
         }
         return objs.toArray(Py.EmptyObjects);
-    }
-
-    /**
-     * Infers the usual Jython executable name from the position of the jar-file returned by
-     * {@link #getJarFileName()} by replacing the file name with "bin/jython". This is intended as
-     * an easy fallback for cases where {@code sys.executable} is {@code None} due to direct
-     * launching via the java executable.
-     * <p>
-     * Note that this does not necessarily return the actual executable, but instead infers the
-     * place where it is usually expected to be. Use {@code sys.executable} to get the actual
-     * executable (may be {@code None}.
-     *
-     * @return usual Jython-executable as absolute path
-     */
-    public static String getDefaultExecutableName() {
-        return getDefaultBinDir() + File.separator
-                + (Platform.IS_WINDOWS ? "jython.exe" : "jython");
-    }
-
-    /**
-     * Infers the usual Jython bin-dir from the position of the jar-file returned by
-     * {@link #getJarFileName()} byr replacing the file name with "bin". This is intended as an easy
-     * fallback for cases where {@code sys.executable} is {@code null} due to direct launching via
-     * the java executable.
-     * <p>
-     * Note that this does not necessarily return the actual bin-directory, but instead infers the
-     * place where it is usually expected to be.
-     *
-     * @return usual Jython bin-dir as absolute path
-     */
-    public static String getDefaultBinDir() {
-        String jar = _getJarFileName();
-        return jar.substring(0, jar.lastIndexOf(File.separatorChar) + 1) + "bin";
-    }
-
-    /**
-     * Utility-method to obtain the name (including absolute path) of the currently used
-     * jython-jar-file. Usually this is jython.jar, but can also be jython-dev.jar or
-     * jython-standalone.jar or something custom.
-     *
-     * @return the full name of the jar file containing this class, <code>null</code> if not
-     *         available.
-     */
-    public static String getJarFileName() {
-        String jar = _getJarFileName();
-        return jar;
-    }
-
-    /**
-     * Utility-method to obtain the name (including absolute path) of the currently used
-     * jython-jar-file. Usually this is jython.jar, but can also be jython-dev.jar or
-     * jython-standalone.jar or something custom.
-     *
-     * @return the full name of the jar file containing this class, <code>null</code> if not
-     *         available.
-     */
-    public static String _getJarFileName() {
-        Class<Py> thisClass = Py.class;
-        String fullClassName = thisClass.getName();
-        String className = fullClassName.substring(fullClassName.lastIndexOf(".") + 1);
-        URL url = thisClass.getResource(className + ".class");
-        return getJarFileNameFromURL(url);
-    }
-
-    /**
-     * Return the path in the file system (as a string) of a JAR located by a URL. Three protocols
-     * are supported, Java JAR-file protocol, and two JBoss protocols "vfs" and "vfszip".
-     * <p>
-     * The JAR-file protocol URL, which must be a {@code jar:file:} reference to a contained element
-     * (that is, it has a "!/" part) is able to identify an actual JAR in a file system that may
-     * then be opened using {@code jarFile = new JarFile(jarFileName)}. The path to the JAR is
-     * returned. If the JAR is accessed by another mechanism ({@code http:} say) this will fail.
-     * <p>
-     * The JBoss URL must be a reference to exactly
-     * {@code vfs:<JAR>/org/python/core/PySystemState.class}, or the same thing using the
-     * {@code vfszip:} protocol, where &lt;JAR&gt; stands for the absolute path to the Jython JAR in
-     * VFS. There is no "!/" marker: in JBoss VFS a JAR is treated just like a directory and can no
-     * longer be opened as a JAR. The method essentially just swaps a VFS protocol for the Java
-     * {@code file:} protocol. The path returned will be correct only if this naive swap is valid.
-     *
-     * @param url into the JAR
-     * @return the file path or {@code null} in the event of a detectable error
-     */
-    public static String getJarFileNameFromURL(URL url) {
-        URI fileURI = null;
-        try {
-            switch (url == null ? "" : url.getProtocol()) {
-
-                case "jar":
-                    // url is jar:file:/some/path/some.jar!/package/with/A.class
-                    if (Platform.IS_WINDOWS) {
-                        // ... or jar:file://host/some/path/some.jar!/package/with/A.class
-                        // ... or jar:file:////host/some/path/some.jar!/package/with/A.class
-                        url = tweakWindowsFileURL(url);
-                    }
-                    URLConnection c = url.openConnection();
-                    fileURI = ((JarURLConnection) c).getJarFileURL().toURI();
-                    break;
-
-                case "vfs":
-                case "vfszip":
-                    // path is /some/path/some-jython.jar/org/python/core/PySystemState.class
-                    String path = url.getPath();
-                    final String target = ".jar/" + Py.class.getName().replace('.', '/');
-                    int jarIndex = path.indexOf(target);
-                    if (jarIndex > 0) {
-                        // path contains the target class in a JAR, so make a file URL for it
-                        fileURI = new URL("file:" + path.substring(0, jarIndex + 4)).toURI();
-                    }
-                    break;
-
-                default:
-                    // Unknown protocol or url==null: fileURI = null
-                    break;
-            }
-        } catch (IOException | URISyntaxException | IllegalArgumentException e) {
-            // Handler cannot open connection or URL is malformed some way: fileURI = null
-        }
-
-        // The JAR file is now identified in fileURI but needs decoding to a file
-        return fileURI == null ? null : new File(fileURI).toString();
-    }
-
-    /**
-     * If the argument is a {@code jar:file:} or {@code file:} URL, compensate for a bug in Java's
-     * construction of URLs affecting {@code java.io.File} and {@code java.net.URLConnection} on
-     * Windows. This is a helper for {@link #getJarFileNameFromURL(URL)}.
-     * <p>
-     * This bug bites when a JAR file is at a (Windows) UNC location, and a {@code jar:file:} URL is
-     * derived from {@code Class.getResource()} as it is in {@link #_getJarFileName()}. When URL is
-     * supplied to {@link #getJarFileNameFromURL(URL)}, the bug leads to a URI that falsely treats a
-     * server as an "authority". It subsequently causes an {@code IllegalArgumentException} with the
-     * message "URI has an authority component" when we try to construct a File. See
-     * {@link https://bugs.java.com/view_bug.do?bug_id=6360233} ("won't fix").
-     *
-     * @param url Possibly malformed URL
-     * @return corrected URL
-     */
-    private static URL tweakWindowsFileURL(URL url) throws MalformedURLException {
-        String urlstr = url.toString();
-        int fileIndex = urlstr.indexOf("file://"); // 7 chars
-        if (fileIndex >= 0) {
-            // Intended UNC path. If there is no slash following these two, insert "/" here:
-            int insert = fileIndex + 7;
-            if (urlstr.length() > insert && urlstr.charAt(insert) != '/') {
-                url = new URL(urlstr.substring(0, insert) + "//" + urlstr.substring(insert));
-            }
-        }
-        return url;
     }
 
 //------------------------constructor-section---------------------------
